improvements on schema of assets, liabilities and recurring rules (generators+templates) and sanitize input

+++ springboot/src/main/java/com/rochanegra/api/vehicles/FuelLogDto.java
@@ -5,18 +5,18 @@ import java.time.LocalDate;
 import java.util.UUID;
 
 public record FuelLogDto(
-                UUID id,
-                UUID vehicleId,
-                BigDecimal quantity,
-                String quantityUnit,
-                BigDecimal pricePerUnit,
-                BigDecimal totalCost,
-                String currency,
-                Integer mileageAtFill,
-                Boolean fullTank,
-                String station,
-                String notes,
-                LocalDate date,
-                BigDecimal normalizedQuantityLiters,
-                Double normalizedMileageKm) {
+        UUID id,
+        UUID vehicleId,
+        BigDecimal quantity,
+        String quantityUnit,
+        BigDecimal pricePerUnit, // <-- This is a calculated field
+        BigDecimal totalCost,
+        String currency,
+        Double mileageAtFill,
+        Boolean fullTank,
+        String station,
+        String notes,
+        LocalDate date,
+        BigDecimal normalizedQuantityLiters,
+        Double normalizedMileageKm) {
 }
